+++ src/main/java/ep/coordinate/CoordinateFactory.java
@@ -1,7 +1,6 @@
 package ep.coordinate;
 
 import ep.coordinate.figure.Figure;
-import ep.coordinate.figure.Vertex;
 import ep.coordinate.figure.factory.FigureFactory;
 import ep.coordinate.figure.factory.RectangleFactory;
 import ep.coordinate.figure.factory.TriangleFactory;
