+++ src/main/java/ep/coordinate/figure/factory/FigureFactory.java
@@ -1,7 +1,7 @@
 package ep.coordinate.figure.factory;
 
 import ep.coordinate.figure.Figure;
-import ep.coordinate.figure.Vertex;
+import ep.coordinate.Vertex;
 
 public interface FigureFactory<T extends Figure> {
