+++ src/test/java/ep/coordinate/TriangleTest.java
@@ -1,7 +1,6 @@
 package ep.coordinate;
 
 import ep.coordinate.figure.Triangle;
-import ep.coordinate.figure.Vertex;
 import ep.coordinate.figure.factory.FigureFactory;
 import ep.coordinate.figure.factory.TriangleFactory;
 import org.junit.jupiter.api.BeforeEach;
